requirements.txt
# ============================================================================
# AuraClass - Sistema Integrado de Gestão Educacional (LMS/ERP)
# Dependências Python para Backend (Django) e Integração IA (AuraMind)
# ============================================================================

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# ============================================================================
# DATABASE
# ============================================================================
psycopg2-binary==2.9.9
django-extensions==3.2.3

# ============================================================================
# PRODUCTION SERVER
# ============================================================================
gunicorn==21.2.0
whitenoise==6.6.0

# ============================================================================
# INTEGRAÇÃO COM AGENTE IA (AuraMind)
# ============================================================================
requests==2.31.0
openai==1.3.8
python-dotenv==1.0.0

# ============================================================================
# MACHINE LEARNING & DATA PROCESSING
# ============================================================================
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# ============================================================================
# NLP & PROCESSAMENTO DE LINGUAGEM NATURAL
# ============================================================================
nltk==3.8.1
spacy==3.7.2

# ============================================================================
# SERIALIZAÇÃO & VALIDAÇÃO
# ============================================================================
marshmallow==3.20.1
pydantic==2.5.0

# ============================================================================
# AUTENTICAÇÃO & SEGURANÇA
# ============================================================================
djangorestframework-simplejwt==5.3.2
django-environ==0.21.0

# ============================================================================
# LOGGING & MONITORAMENTO
# ============================================================================
python-json-logger==2.0.7
sentry-sdk==1.39.1

# ============================================================================
# TESTES
# ============================================================================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# ============================================================================
# DESENVOLVIMENTO
# ============================================================================
black==23.12.0
flake8==6.1.0
isort==5.13.2
pre-commit==3.5.0

# ============================================================================
# DOCUMENTAÇÃO
# ============================================================================
drf-spectacular==0.26.5
drf-spectacular-sidecar==0.26.5

# ============================================================================
# UTILITÁRIOS
# ============================================================================
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
